requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-experimental==0.0.64
langchain-core==0.2.38
langchain-chroma
chromadb==0.5.23
openai==1.55.3
httpx==0.27.2
pandas==2.2.2
python-dotenv==1.0.1
cos-python-sdk-v5==1.9.30
tiktoken==0.7.0
pypdf==6.8.0
beautifulsoup4==4.14.3
soupsieve==2.8.3
